requirements.txt
asyncio
discord
python-dotenv
requests
sentencepiece